test(integration): extract IntegrationTestBase
Pulls the container/lifecycle/helper scaffolding out of
ScimPropagationFromLdapIT into an abstract IntegrationTestBase that
future test classes can extend. The behavior of every existing test
is unchanged — verified by running the full suite (4 + 12 = 16 tests,
all green).

What moved to the base class:

- Static container fields (KeycloakContainer, openldap, network,
  WireMock, admin) plus their @BeforeAll/@AfterAll lifecycle.
- @BeforeEach stub reset.
- TestRealm record.
- Realm/component setup helpers: newRealmWithScimAndLdap (+ config
  customizer flavor), addScimStorageProvider, addLdapFederation,
  addLdapAttributeMapper, attachScimMapper, createAdminUser,
  enableScimEventListener.
- WireMock stub helpers: stubScimUserCreateOk, stubScimUserUpdateOk,
  stubScimUserDeleteOk, awaitUserPostFor.
- LDAP manipulation: newLdapEnv, deleteLdapEntry, modifyLdapAttribute,
  reAddAlice.
- Reconciler endpoint: postReconcile.
- Generic: sleepQuietly.

Naming notes:
- Stub helpers gained a 'User' segment (stubScimUserCreateOk vs the
  former stubScimCreateOk) to disambiguate from the Group stubs the
  next commit will add.
- awaitPostFor renamed to awaitUserPostFor for the same reason.

Forked-JVM mechanics (forkEvery=1) keep this safe: each subclass gets
its own JVM, so static container fields don't leak between classes.
The base class's @BeforeAll runs first per JUnit 5's superclass-first
ordering; subclasses get a fully-initialized environment.
